*: move "alarm,compactor" to "internal/*"
