+++ internal/compactor/compactor_test.go
