+++ internal/compactor/doc.go
