+++ internal/compactor/periodic.go
