+++ internal/compactor/periodic_test.go
